Add the functionality for displaying week and month productivity charts apart from the day charts

+++ app/src/main/java/com/jdevs/timeo/model/RecordItem.kt
@@ -7,7 +7,7 @@ import org.threeten.bp.OffsetDateTime
 data class RecordItem(
     override val id: String,
     val name: String,
-    val time: Long,
+    val time: Int,
     val activityId: String,
     val creationDate: OffsetDateTime
 ) : ViewItem {
