+++ app/src/main/java/com/jdevs/timeo/ui/common/adapter/LoadingDelegateAdapter.kt
@@ -10,7 +10,7 @@ class LoadingDelegateAdapter : DelegateAdapter {
 
     override fun onCreateViewHolder(
         parent: ViewGroup,
-        createRecord: (Int, Long) -> Unit,
+        createRecord: (Int, Int) -> Unit,
         navigateToDetails: (Int) -> Unit,
         showDeleteDialog: (Int) -> Unit
     ) = ViewHolder(parent)
