+++ app/src/main/java/com/jdevs/timeo/ui/projects/ProjectDelegateAdapter.kt
@@ -17,7 +17,7 @@ class ProjectDelegateAdapter : DelegateAdapter {
 
     override fun onCreateViewHolder(
         parent: ViewGroup,
-        createRecord: (Int, Long) -> Unit,
+        createRecord: (Int, Int) -> Unit,
         navigateToDetails: (Int) -> Unit,
         showDeleteDialog: (Int) -> Unit
     ): RecyclerView.ViewHolder {
@@ -44,7 +44,7 @@ class ProjectDelegateAdapter : DelegateAdapter {
     class ViewHolder(
         rootView: View,
         private val viewModel: ProjectViewModel,
-        private val createRecord: (Int, Long) -> Unit = { _, _ -> },
+        private val createRecord: (Int, Int) -> Unit = { _, _ -> },
         private val navigateToDetails: (Int) -> Unit = {}
     ) : PagingAdapter.ViewHolder(rootView) {
