+++ app/src/main/java/com/jdevs/timeo/util/time/Converters.kt
@@ -2,4 +2,4 @@ package com.jdevs.timeo.util.time
 
 import com.jdevs.timeo.shared.time.HOUR_MINUTES
 
-fun getMins(hours: Long, minutes: Long) = hours * HOUR_MINUTES + minutes
+fun getMins(hours: Int, minutes: Int) = hours * HOUR_MINUTES + minutes
