+++ data/src/main/java/com/jdevs/timeo/data/stats/Stats.kt
@@ -14,18 +14,18 @@ abstract class FirestoreStats {
     val documentId = ""
 }
 
-data class FirestoreDayStats(val time: Long = 0, val day: Long = 0) : FirestoreStats()
-data class FirestoreWeekStats(val time: Long = 0, val day: Int = 0) : FirestoreStats()
-data class FirestoreMonthStats(val time: Long = 0, val day: Int = 0) : FirestoreStats()
+data class FirestoreDayStats(val time: Int = 0, val day: Int = 0) : FirestoreStats()
+data class FirestoreWeekStats(val time: Int = 0, val day: Int = 0) : FirestoreStats()
+data class FirestoreMonthStats(val time: Int = 0, val day: Int = 0) : FirestoreStats()
 
 @Entity(tableName = "dayStats")
-data class DBDayStats(@PrimaryKey val day: Long = 0, val time: Long = 0)
+data class DBDayStats(@PrimaryKey val day: Int = 0, val time: Int = 0)
 
 @Entity(tableName = "weekStats")
-data class DBWeekStats(@PrimaryKey val week: Int = 0, val time: Long = 0)
+data class DBWeekStats(@PrimaryKey val week: Int = 0, val time: Int = 0)
 
 @Entity(tableName = "monthStats")
-data class DBMonthStats(@PrimaryKey val month: Int = 0, val time: Long = 0)
+data class DBMonthStats(@PrimaryKey val month: Int = 0, val time: Int = 0)
 
 fun DBDayStats.mapToDomain() = DayStats(day.toString(), time, day)
 fun DBWeekStats.mapToDomain() = WeekStats(week.toString(), time, week)
