+++ domain/src/main/java/com/jdevs/timeo/domain/model/Record.kt
@@ -5,7 +5,7 @@ import org.threeten.bp.OffsetDateTime
 data class Record(
     val id: String = "",
     val name: String,
-    val time: Long,
+    val time: Int,
     val activityId: String = "",
     val creationDate: OffsetDateTime = OffsetDateTime.now()
 )
