+++ domain/src/main/java/com/jdevs/timeo/domain/model/Stats.kt
@@ -1,5 +1,5 @@
 package com.jdevs.timeo.domain.model
 
-data class DayStats(val id: String = "", val time: Long, val day: Long)
-data class WeekStats(val id: String = "", val time: Long, val week: Int)
-data class MonthStats(val id: String = "", val time: Long, val month: Int)
+data class DayStats(val id: String = "", val time: Int, val day: Int)
+data class WeekStats(val id: String = "", val time: Int, val week: Int)
+data class MonthStats(val id: String = "", val time: Int, val month: Int)
